# src/PaperlessREST/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/token" : {
      "post" : {
        "operationId" : "GetToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      }
    },
    "/api/groups" : {
      "get" : {
        "operationId" : "GetGroups",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "page_size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":0,\"next\":null,\"previous\":null,\"all\":[],\"results\":[]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetGroups_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      },
      "post" : {
        "operationId" : "CreateGroup",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"name\":\"testgroup\",\"permissions\":[\"add_document\",\"view_document\",\"change_document\",\"delete_document\",\"add_tag\",\"view_tag\",\"change_tag\",\"delete_tag\",\"add_correspondent\",\"view_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"add_documenttype\",\"view_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"add_storagepath\",\"view_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"add_savedview\",\"view_savedview\",\"change_savedview\",\"delete_savedview\",\"add_paperlesstask\",\"view_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"add_uisettings\",\"view_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"add_note\",\"view_note\",\"change_note\",\"delete_note\",\"add_mailaccount\",\"view_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"add_mailrule\",\"view_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"add_user\",\"view_user\",\"change_user\",\"delete_user\",\"add_group\",\"view_group\",\"change_group\",\"delete_group\",\"add_logentry\",\"view_logentry\",\"change_logentry\",\"delete_logentry\"]}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/CreateGroup_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":1,\"name\":\"testgroup\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}"
                  }
                },
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      }
    },
    "/api/groups/{id}" : {
      "delete" : {
        "operationId" : "DeleteGroup",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      },
      "put" : {
        "operationId" : "UpdateGroup",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"id\":1,\"name\":\"testgroup2\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"],\"set_permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"],\"permissions_form\":{\"set_permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/UpdateGroup_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"id\":1,\"name\":\"testgroup2\",\"permissions\":[\"add_logentry\",\"change_logentry\",\"delete_logentry\",\"view_logentry\",\"add_group\",\"change_group\",\"delete_group\",\"view_group\",\"add_user\",\"change_user\",\"delete_user\",\"view_user\",\"add_correspondent\",\"change_correspondent\",\"delete_correspondent\",\"view_correspondent\",\"add_document\",\"change_document\",\"delete_document\",\"view_document\",\"add_documenttype\",\"change_documenttype\",\"delete_documenttype\",\"view_documenttype\",\"add_note\",\"change_note\",\"delete_note\",\"view_note\",\"add_paperlesstask\",\"change_paperlesstask\",\"delete_paperlesstask\",\"view_paperlesstask\",\"add_savedview\",\"change_savedview\",\"delete_savedview\",\"view_savedview\",\"add_storagepath\",\"change_storagepath\",\"delete_storagepath\",\"view_storagepath\",\"add_tag\",\"change_tag\",\"delete_tag\",\"view_tag\",\"add_uisettings\",\"change_uisettings\",\"delete_uisettings\",\"view_uisettings\",\"add_mailaccount\",\"change_mailaccount\",\"delete_mailaccount\",\"view_mailaccount\",\"add_mailrule\",\"change_mailrule\",\"delete_mailrule\",\"view_mailrule\"]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/UpdateGroup_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      }
    },
...
    "/api/logs/{id}" : {
      "get" : {
        "operationId" : "GetLog",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "[\"[2023-09-08 12:03:51,945] [INFO] [paperless.management.consumer] Using inotify to watch directory for changes: /usr/src/paperless/consume\",\"[2023-09-08 12:04:53,939] [INFO] [paperless.consumer] Consuming test.pdf\",\"[2023-09-08 12:04:53,947] [DEBUG] [paperless.consumer] Detected mime type: application/pdf\",\"[2023-09-08 12:04:54,030] [DEBUG] [paperless.consumer] Parser: RasterisedDocumentParser\",\"[2023-09-08 12:04:54,036] [DEBUG] [paperless.consumer] Parsing test.pdf...\",\"[2023-09-08 12:04:55,095] [DEBUG] [paperless.parsing.tesseract] Calling OCRmyPDF with args: {'input_file': PosixPath('/tmp/paperless/paperless-ngxm31l7wip/test.pdf'), 'output_file': PosixPath('/tmp/paperless/paperless-iabw4ryn/archive.pdf'), 'use_threads': True, 'jobs': '1', 'language': 'eng', 'output_type': 'pdfa', 'progress_bar': False, 'skip_text': True, 'clean': True, 'deskew': True, 'rotate_pages': True, 'rotate_pages_threshold': 12.0, 'sidecar': PosixPath('/tmp/paperless/paperless-iabw4ryn/sidecar.txt')}\",\"[2023-09-08 12:04:59,229] [DEBUG] [paperless.parsing.tesseract] Incomplete sidecar file: discarding.\",\"[2023-09-08 12:04:59,460] [DEBUG] [paperless.consumer] Generating thumbnail for test.pdf...\",\"[2023-09-08 12:04:59,561] [DEBUG] [paperless.parsing] Execute: convert -density 300 -scale 500x5000> -alpha remove -strip -auto-orient /tmp/paperless/paperless-iabw4ryn/archive.pdf[0] /tmp/paperless/paperless-iabw4ryn/convert.webp\",\"[2023-09-08 12:05:27,938] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:05:28,195] [DEBUG] [paperless.consumer] Saving record to database\",\"[2023-09-08 12:05:28,258] [DEBUG] [paperless.consumer] Creation date from st_mtime: 2023-09-08 12:04:53+00:00\",\"[2023-09-08 12:05:29,748] [DEBUG] [paperless.consumer] Deleting file /tmp/paperless/paperless-ngxm31l7wip/test.pdf\",\"[2023-09-08 12:05:29,832] [DEBUG] [paperless.parsing.tesseract] Deleting directory /tmp/paperless/paperless-iabw4ryn\",\"[2023-09-08 12:05:29,836] [INFO] [paperless.consumer] Document 2023-09-08 test consumption finished\",\"[2023-09-08 12:05:37,518] [INFO] [paperless.tasks] No automatic matching items, not training\",\"[2023-09-08 12:06:21,057] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:21,069] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:22,956] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:06:24,146] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:27,642] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:27,852] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:27,874] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:10:36,578] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:43,556] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:52,429] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:10:57,501] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:11:10,623] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:46,620] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:47,230] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:14:55,240] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:01,444] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:02,965] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:15,266] [DEBUG] [paperless.classifier] Document classification model does not exist (yet), not performing automatic matching.\",\"[2023-09-08 12:15:24,462] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\",\"[2023-09-08 12:15:24,661] [DEBUG] [paperless.filehandling] Document has storage_path 1 (testdirp) set\"]"
                  }
                },
                "schema" : {
                  "items" : {
                    "type" : "string"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Config" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserInfo" : {
        "additionalProperties" : false,
        "example" : {
          "password" : "password",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "nullable" : true,
            "type" : "string"
          },
          "password" : {
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GetGroups_200_response" : {
        "example" : {
          "next" : 6,
          "all" : [ "", "" ],
          "previous" : 1,
          "count" : 0,
          "results" : [ "", "" ]
        },
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "items" : { },
            "type" : "array"
          },
          "results" : {
            "items" : { },
            "type" : "array"
          }
        },
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object"
      },
      "CreateGroup_request" : {
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "properties" : {
                "name" : {
                  "type" : "string"
                },
                "permissions" : {
                  "items" : {
                    "type" : "string"
                  },
                  "type" : "array"
                }
              },
              "required" : [ "name", "permissions" ],
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "id", "name", "permissions" ],
        "type" : "object"
      },
      "UpdateGroup_request_permissions_form" : {
        "properties" : {
          "set_permissions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "set_permissions" ],
        "type" : "object"
      },
      "UpdateGroup_request" : {
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "set_permissions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "permissions_form" : {
            "$ref" : "#/components/schemas/UpdateGroup_request_permissions_form"
          }
        },
        "required" : [ "id", "name", "permissions", "permissions_form", "set_permissions" ],
        "type" : "object"
      },
      "UpdateGroup_200_response" : {
        "example" : {
          "permissions" : [ "permissions", "permissions" ],
          "name" : "name",
          "id" : 0
        },
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "id", "name", "permissions" ],
        "type" : "object"
      },
...
    }
  }
}